+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 - Add `before_send_check_in` for applying to `CheckInEvent` ([#2703](https://github.com/getsentry/sentry-ruby/pull/2703))
 - Returning a hash from `before_send` and `before_send_transaction` is no longer supported and will drop the event.
 
+### Internal
+
+- Archive [`sentry-raven`](https://github.com/getsentry/raven-ruby) ([#2708](https://github.com/getsentry/sentry-ruby/pull/2708))
+
 ## Unreleased
 
 ### Bug Fixes