+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Bug Fixes
 
 - Skip including `sentry.message.template` in the log event attributes if there are no interpolation parameters provided ([#2700](https://github.com/getsentry/sentry-ruby/pull/2700))
+- Respect `log_level` when logging via `:std_lib_logger` patch ([#2709](https://github.com/getsentry/sentry-ruby/pull/2709))
 
 ## 5.27.0